Change points type to BigInt and open Dice game for all

+++ src/bot/subscription.ts
@@ -72,7 +72,7 @@ export class Subscription {
     }
 
     // amount in nano-Toncoins (1 Toncoin = 1e9 nano-Toncoins)
-    const points = Math.round(Number(ton) * 100_000);
+    const points = BigInt(Math.round(Number(ton))) * BigInt(100_000);
     logger.debug(`${ton} => ${points}`);
     user.votes += points;
     await user.save();
